Fix crash on out of range MIDI events (though this shouldn't be possible at all....).
git-svn-id: svn://localhost/ardour2/branches/3.0@3357 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/midi_region_view.cc
@@ -764,8 +764,8 @@ MidiRegionView::add_note(const boost::shared_ptr<Note> note)
 		ev_rect->property_y2() = y1 + floor(midi_stream_view()->note_height());
 
 		if (note->duration() == 0) {
-			cerr << "MidiModel: WARNING: Discovered note with duration 0 and pitch" << note->note() 
-			     << " at time " << note->time() << endl;
+			cerr << "MidiModel: WARNING: Discovered note with duration 0 and pitch"
+				<< (int)note->note() << " at time " << note->time() << endl;
 			if (_active_notes) {
 				assert(note->note() < 128);
 				_active_notes[note->note()] = ev_rect;
